Rename `es` to `elasticsearch`
