+++ elasticsearch/src/main/scala/monix/connect/elasticsearch/package.scala
@@ -21,12 +21,12 @@ import com.sksamuel.elastic4s._
 import monix.eval.Task
 import monix.execution.internal.InternalApi
 
-package object es {
+package object elasticsearch {
 
-  @InternalApi private[es] implicit val taskFunctor: Functor[Task] = new Functor[Task] {
+  @InternalApi private[elasticsearch] implicit val taskFunctor: Functor[Task] = new Functor[Task] {
     override def map[A, B](fa: Task[A])(f: A => B): Task[B] = fa.map(f)
   }
 
-  @InternalApi private[es] implicit val taskExecutor: Executor[Task] = (client: HttpClient, request: ElasticRequest) =>
-    Task.async(k => client.send(request, k))
+  @InternalApi private[elasticsearch] implicit val taskExecutor: Executor[Task] =
+    (client: HttpClient, request: ElasticRequest) => Task.async(k => client.send(request, k))
 }
